test: split kv tests out into their own file
The key-value functionality in wrangler is pretty extensive and is likely to have a lot of tests. So this commit simply moves it out into its own file to make things a bit more manageable.
