+++ packages/wrangler/src/__tests__/index.test.ts
@@ -3,34 +3,13 @@ import * as fsp from "node:fs/promises";
 import * as os from "node:os";
 import * as path from "node:path";
 import * as TOML from "@iarna/toml";
-import { main } from "../index";
-import { setMock, unsetAllMocks } from "./mock-cfetch";
 import { mockConfirm } from "./mock-dialogs";
-
-jest.mock("../cfetch", () => jest.requireActual("./mock-cfetch"));
-
-async function w(cmd?: string) {
-  const logSpy = jest.spyOn(console, "log").mockImplementation();
-  const errorSpy = jest.spyOn(console, "error").mockImplementation();
-  const warnSpy = jest.spyOn(console, "warn").mockImplementation();
-  try {
-    await main(cmd?.split(" ") ?? []);
-    return {
-      stdout: logSpy.mock.calls.flat(2).join("\n"),
-      stderr: errorSpy.mock.calls.flat(2).join("\n"),
-      warnings: warnSpy.mock.calls.flat(2).join("\n"),
-    };
-  } finally {
-    logSpy.mockRestore();
-    errorSpy.mockRestore();
-    warnSpy.mockRestore();
-  }
-}
+import { runWrangler } from "./run-wrangler";
 
 describe("wrangler", () => {
   describe("no command", () => {
     it("should display a list of available commands", async () => {
-      const { stdout, stderr } = await w();
+      const { stdout, stderr } = await runWrangler();
 
       expect(stdout).toMatchInlineSnapshot(`
         "wrangler
@@ -61,7 +40,7 @@ describe("wrangler", () => {
 
   describe("invalid command", () => {
     it("should display an error", async () => {
-      const { stdout, stderr } = await w("invalid-command");
+      const { stdout, stderr } = await runWrangler("invalid-command");
 
       expect(stdout).toMatchInlineSnapshot(`
         "wrangler
@@ -112,7 +91,7 @@ describe("wrangler", () => {
         text: "No package.json found. Would you like to create one?",
         result: false,
       });
-      await w("init");
+      await runWrangler("init");
       const parsed = TOML.parse(await fsp.readFile("./wrangler.toml", "utf-8"));
       expect(typeof parsed.compatibility_date).toBe("string");
       expect(fs.existsSync("./package.json")).toBe(false);
@@ -125,7 +104,7 @@ describe("wrangler", () => {
         text: "Do you want to continue initializing this project?",
         result: false,
       });
-      const { stderr } = await w("init");
+      const { stderr } = await runWrangler("init");
       expect(stderr).toContain("wrangler.toml file already exists!");
       const parsed = TOML.parse(await fsp.readFile("./wrangler.toml", "utf-8"));
       expect(typeof parsed.compatibility_date).toBe("undefined");
@@ -143,7 +122,7 @@ describe("wrangler", () => {
           result: false,
         }
       );
-      const { stderr } = await w("init");
+      const { stderr } = await runWrangler("init");
       expect(stderr).toContain("wrangler.toml file already exists!");
       const parsed = TOML.parse(await fsp.readFile("./wrangler.toml", "utf-8"));
       expect(typeof parsed.compatibility_date).toBe("string");
@@ -160,7 +139,7 @@ describe("wrangler", () => {
           result: false,
         }
       );
-      await w("init");
+      await runWrangler("init");
       expect(fs.existsSync("./package.json")).toBe(true);
       const packageJson = JSON.parse(
         fs.readFileSync("./package.json", "utf-8")
@@ -182,7 +161,7 @@ describe("wrangler", () => {
         "utf-8"
       );
 
-      await w("init");
+      await runWrangler("init");
       const packageJson = JSON.parse(
         fs.readFileSync("./package.json", "utf-8")
       );
@@ -205,7 +184,7 @@ describe("wrangler", () => {
       fs.mkdirSync("./sub-1/sub-2", { recursive: true });
       process.chdir("./sub-1/sub-2");
 
-      await w("init");
+      await runWrangler("init");
       expect(fs.existsSync("./package.json")).toBe(false);
       expect(fs.existsSync("../../package.json")).toBe(true);
 
@@ -227,7 +206,7 @@ describe("wrangler", () => {
           result: true,
         }
       );
-      await w("init");
+      await runWrangler("init");
       expect(fs.existsSync("./tsconfig.json")).toBe(true);
       const tsconfigJson = JSON.parse(
         fs.readFileSync("./tsconfig.json", "utf-8")
@@ -255,7 +234,7 @@ describe("wrangler", () => {
         "utf-8"
       );
 
-      await w("init");
+      await runWrangler("init");
       const tsconfigJson = JSON.parse(
         fs.readFileSync("./tsconfig.json", "utf-8")
       );
@@ -277,7 +256,7 @@ describe("wrangler", () => {
       fs.mkdirSync("./sub-1/sub-2", { recursive: true });
       process.chdir("./sub-1/sub-2");
 
-      await w("init");
+      await runWrangler("init");
       expect(fs.existsSync("./tsconfig.json")).toBe(false);
       expect(fs.existsSync("../../tsconfig.json")).toBe(true);
 
@@ -288,104 +267,32 @@ describe("wrangler", () => {
     });
 
     it("should error if `--type` is used", async () => {
-      const noValue = await w("init --type");
+      const noValue = await runWrangler("init --type");
       expect(noValue.stderr).toMatchInlineSnapshot(
         `"The --type option is no longer supported."`
       );
     });
 
     it("should error if `--type javascript` is used", async () => {
-      const javascriptValue = await w("init --type javascript");
+      const javascriptValue = await runWrangler("init --type javascript");
       expect(javascriptValue.stderr).toMatchInlineSnapshot(
         `"The --type option is no longer supported."`
       );
     });
 
     it("should error if `--type rust` is used", async () => {
-      const rustValue = await w("init --type rust");
+      const rustValue = await runWrangler("init --type rust");
       expect(rustValue.stderr).toMatchInlineSnapshot(
         `"The --type option is no longer supported."`
       );
     });
 
     it("should error if `--type webpack` is used", async () => {
-      const webpackValue = await w("init --type webpack");
+      const webpackValue = await runWrangler("init --type webpack");
       expect(webpackValue.stderr).toMatchInlineSnapshot(`
         "The --type option is no longer supported.
         If you wish to use webpack then you will need to create a custom build."
       `);
-    });
-  });
-
-  describe("kv:namespace", () => {
-    afterEach(() => {
-      unsetAllMocks();
-    });
-
-    it("can create a namespace", async () => {
-      const KVNamespaces: { title: string; id: string }[] = [];
-      setMock("/accounts/:accountId/storage/kv/namespaces", (uri, init) => {
-        expect(init.method === "POST");
-        expect(uri[0]).toEqual(
-          "/accounts/some-account-id/storage/kv/namespaces"
-        );
-        const { title } = JSON.parse(init.body);
-        expect(title).toEqual("worker-UnitTestNamespace");
-        KVNamespaces.push({ title, id: "some-namespace-id" });
-        return { id: "some-namespace-id" };
-      });
-
-      await w("kv:namespace create UnitTestNamespace");
-
-      expect(KVNamespaces).toEqual([
-        {
-          title: "worker-UnitTestNamespace",
-          id: "some-namespace-id",
-        },
-      ]);
-    });
-
-    it("can list namespaces", async () => {
-      const KVNamespaces: { title: string; id: string }[] = [
-        { title: "title-1", id: "id-1" },
-        { title: "title-2", id: "id-2" },
-      ];
-      setMock(
-        "/accounts/:accountId/storage/kv/namespaces\\?:qs",
-        (uri, init) => {
-          expect(uri[0]).toContain(
-            "/accounts/some-account-id/storage/kv/namespaces"
-          );
-          expect(uri[2]).toContain("per_page=100");
-          expect(uri[2]).toContain("order=title");
-          expect(uri[2]).toContain("direction=asc");
-          expect(uri[2]).toContain("page=1");
-          expect(init).toBe(undefined);
-          return KVNamespaces;
-        }
-      );
-      const { stdout } = await w("kv:namespace list");
-      const namespaces = JSON.parse(stdout) as { id: string; title: string }[];
-      expect(namespaces).toEqual(KVNamespaces);
-    });
-
-    it("can delete a namespace", async () => {
-      let accountId = "";
-      let namespaceId = "";
-      setMock(
-        "/accounts/:accountId/storage/kv/namespaces/:namespaceId",
-        (uri, init) => {
-          accountId = uri[1];
-          namespaceId = uri[2];
-          expect(uri[0]).toEqual(
-            "/accounts/some-account-id/storage/kv/namespaces/some-namespace-id"
-          );
-          expect(init.method).toBe("DELETE");
-        }
-      );
-      await w(`kv:namespace delete --namespace-id some-namespace-id`);
-      expect(accountId).toEqual("some-account-id");
-      expect(namespaceId).toEqual("some-namespace-id");
     });
   });
 });
